requirements.txt
matplotlib>=3.5.0
numpy>=1.20.0
scipy>=1.7.0
torch>=1.10.0
torchvision>=0.11.0
tqdm>=4.62.0
trimesh>=3.9.0
numpy-stl>=2.17.1
